Make GraphiQL more conventional in dev2
